third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/svg/coordinate-systems/svgtransformlist-replaceitem-expected.txt

This is a testharness.js-based test.
FAIL Dynamic update of transform; replaceItem() assert_equals: expected 185.640625 but got 185
Harness: the test ran to completion.
